Add GPU vertex/index buffer binding and depth functionality (#85)
This change adds all the missing bindings to the GPU module for
creating, mapping, and uploading vertex and index buffers, and also
fixes the graphics pipeline creation bindings so that it's possible to
set up a depth buffer for full 3D graphics.

It also adds a new `gpu-cube` example to demonstrate this functionality.

There's also a minor bug fix in here where the `FillMode` being set on
the graphics pipeline wasn't applying because it was being lost in the
pipeline creation function.